fix: include custom-keys command into enroll-keys cmd
